Optimize dependency declaration
Extract declaring dependency versions required for upper bound management
with the maven-enforcer-plugin to the `<dependencyManagement>` element.
This just defines the versions but not the actual dependencies. Only
direct dependencies should and are now declared in the `<dependencies>`
element. This provides a better overview what is actually required.
It also significantly reduces the size of the HPI file as less JAR files
are packaged now.