Optimize dependency declaration (#67)
* Optimize dependency declaration

Extract declaring dependency versions required for upper bound management
with the maven-enforcer-plugin to the `<dependencyManagement>` element.
This just defines the versions but not the actual dependencies. Only
direct dependencies should and are now declared in the `<dependencies>`
element. This provides a better overview what is actually required.
It also significantly reduces the size of the HPI file as less JAR files
are packaged now.

* Overwrite jackson-databind version

It was upgraded in commit f689b2e3b96efdae3472b411f32b2d8af6763ac7
to get the latest security patches.

Even not beeing a direct but a transitive dependency, overwrite
its version for security reasons as stated above.